refactoring test layout and cmake logic
